+++ e2e/commands/init.e2e.1.js
@@ -6,8 +6,9 @@ import { BIT_GIT_DIR, BIT_HIDDEN_DIR, BIT_MAP, BIT_JSON } from '../../src/consta
 // import bitImportGitHook from '../../src/git-hooks/fixtures/bit-import-git-hook';
 import { ScopeJsonNotFound } from '../../src/scope/exceptions';
 import { InvalidBitMap } from '../../src/consumer/bit-map/exceptions';
-import { InvalidBitConfig } from '../../src/consumer/bit-config/exceptions';
+import { InvalidBitJson } from '../../src/consumer/bit-config/exceptions';
 import { statusWorkspaceIsCleanMsg } from '../../src/cli/commands/public-cmds/status-cmd';
+import InvalidPackageJson from '../../src/consumer/bit-config/exceptions/invalid-package-json';
 
 const assertArrays = require('chai-arrays');
 
@@ -198,10 +199,10 @@ describe('run bit init', function () {
         helper.reInitLocalScope();
         helper.corruptBitJson();
       });
-      it('bit status should throw an exception InvalidBitConfig', () => {
+      it('bit status should throw an exception InvalidBitJson', () => {
         const bitJsonPath = path.join(helper.localScopePath, BIT_JSON);
         const statusCmd = () => helper.runCmd('bit status');
-        const error = new InvalidBitConfig(bitJsonPath, 'Unexpected token t');
+        const error = new InvalidBitJson(bitJsonPath, 'Unexpected token t');
         helper.expectToThrow(statusCmd, error);
       });
       it('should create a new bit.json file', () => {
@@ -299,6 +300,51 @@ describe('run bit init', function () {
       it('bit status should show nothing-to-tag', () => {
         const output = helper.runCmd('bit status');
         expect(output).to.have.string(statusWorkspaceIsCleanMsg);
+      });
+    });
+  });
+  describe('when a project has package.json file', () => {
+    describe('without --standalone flag', () => {
+      before(() => {
+        helper.cleanLocalScope();
+        helper.initNpm();
+        helper.runCmd('bit init');
+      });
+      it('should write the bit.json content into the package.json inside "bit" property', () => {
+        const packageJson = helper.readPackageJson();
+        expect(packageJson).to.have.property('bit');
+        expect(packageJson.bit).to.have.property('componentsDefaultDirectory');
+        expect(packageJson.bit.componentsDefaultDirectory).to.equal('components/{name}');
+      });
+      it('should not create bit.json file', () => {
+        expect(path.join(helper.localScopePath, 'bit.json')).to.not.be.a.path();
+      });
+    });
+    describe('with --standalone flag', () => {
+      before(() => {
+        helper.cleanLocalScope();
+        helper.initNpm();
+        helper.runCmd('bit init --standalone');
+      });
+      it('should not write the bit.json content into the package.json file', () => {
+        const packageJson = helper.readPackageJson();
+        expect(packageJson).to.not.have.property('bit');
+      });
+      it('should create bit.json file', () => {
+        expect(path.join(helper.localScopePath, 'bit.json')).to.be.a.file();
+        const bitJson = helper.readBitJson();
+        expect(bitJson).to.have.property('componentsDefaultDirectory');
+      });
+    });
+    describe('when the package.json is corrupted', () => {
+      before(() => {
+        helper.cleanLocalScope();
+        helper.corruptPackageJson();
+      });
+      it('should throw InvalidPackageJson error', () => {
+        const initCmd = () => helper.runCmd('bit init');
+        const error = new InvalidPackageJson(path.join(helper.localScopePath, 'package.json'));
+        helper.expectToThrow(initCmd, error);
       });
     });
   });