+++ src/cli/default-error-handler.js
@@ -45,7 +45,8 @@ import {
   CyclicDependencies,
   HashNotFound
 } from '../scope/exceptions';
-import InvalidBitConfig from '../consumer/bit-config/exceptions/invalid-bit-config';
+import InvalidBitJson from '../consumer/bit-config/exceptions/invalid-bit-json';
+import InvalidPackageJson from '../consumer/bit-config/exceptions/invalid-package-json';
 import InvalidVersion from '../api/consumer/lib/exceptions/invalid-version';
 import NoIdMatchWildcard from '../api/consumer/lib/exceptions/no-id-match-wildcard';
 import NothingToCompareTo from '../api/consumer/lib/exceptions/nothing-to-compare-to';
@@ -303,15 +304,20 @@ to re-start Bit from scratch, deleting all objects from the scope, use "bit init
       )}" is invalid, component scope names can only contain alphanumeric, lowercase characters, and the following ["-", "_", "$", "!"]`
   ],
   [
-    InvalidBitConfig,
+    InvalidBitJson,
     err => `error: invalid bit.json: ${chalk.bold(err.path)} is not a valid JSON file.
-    consider running ${chalk.bold('bit init --reset')} to recreate the file`
+consider running ${chalk.bold('bit init --reset')} to recreate the file`
+  ],
+  [
+    InvalidPackageJson,
+    err => `error: package.json at ${chalk.bold(err.path)} is not a valid JSON file.
+please fix the file in order to run bit commands`
   ],
   [
     InvalidBitConfigPropPath,
     err => `error: the path "${chalk.bold(err.fieldValue)}" of "${chalk.bold(
       err.fieldName
-    )}" in your bit.json file is invalid.
+    )}" in your bit.json or package.json file is invalid.
 please make sure it's not absolute and doesn't contain invalid characters`
   ],
   [