+++ src/consumer/bit-config/exceptions/index.js
@@ -1,6 +1,7 @@
 /** @flow */
-import BitJsonNotFound from './bit-config-not-found';
-import InvalidBitConfig from './invalid-bit-config';
+import BitConfigNotFound from './bit-config-not-found';
+import InvalidBitJson from './invalid-bit-json';
+import InvalidPackageJson from './invalid-package-json';
 import InvalidBitConfigPropPath from './invalid-bit-config-prop-path';
 
-export { BitJsonNotFound, InvalidBitConfig, InvalidBitConfigPropPath };
+export { BitConfigNotFound, InvalidBitJson, InvalidPackageJson, InvalidBitConfigPropPath };