+++ src/consumer/component-ops/eject-conf.js
@@ -15,7 +15,6 @@ import TesterExtension from '../../extensions/tester-extension';
 import type { PathOsBased } from '../../utils/path';
 import DataToPersist from '../component/sources/data-to-persist';
 import { COMPILER_ENV_TYPE, TESTER_ENV_TYPE } from '../../constants';
-import JSONFile from '../component/sources/json-file';
 import RemovePath from '../component/sources/remove-path';
 import AbstractBitConfig from '../bit-config/abstract-bit-config';
 
@@ -108,24 +107,15 @@ export async function getEjectConfDataToPersist(
     relativeEjectedCompilerDirectory,
     relativeEjectedTesterDirectory
   );
-  const bitJsonDataToWrite = await bitJson.prepareToWrite({ bitDir: bitJsonDir.dirPath });
-  if (bitJsonDataToWrite) {
-    dataToPersist.addFile(
-      JSONFile.load({
-        base: bitJsonDir.dirPath,
-        path: bitJsonDataToWrite.pathToWrite,
-        override: true,
-        content: bitJsonDataToWrite.content
-      })
-    );
-  }
+  const jsonFilesToWrite = await bitJson.prepareToWrite({ bitDir: bitJsonDir.dirPath });
+  dataToPersist.addManyFiles(jsonFilesToWrite);
 
   if (deleteOldFiles) {
     if (oldConfigDir) {
       const oldBitJsonDir = oldConfigDir.getResolved({ componentDir }).getEnvTypeCleaned();
       const oldBitJsonDirFullPath = path.join(consumerPath, oldBitJsonDir.dirPath);
       if (bitJsonDirFullPath !== oldBitJsonDirFullPath) {
-        const bitJsonToRemove = AbstractBitConfig.composePath(oldBitJsonDir.dirPath);
+        const bitJsonToRemove = AbstractBitConfig.composeBitJsonPath(oldBitJsonDir.dirPath);
         dataToPersist.removePath(new RemovePath(bitJsonToRemove, true));
       }
     }